Reorganize classes and remove unused stuff
