+++ src/main/scala/nl/codestar/util/Distance.scala
@@ -1,6 +1,6 @@
 package nl.codestar.util
 
-import nl.codestar.data.Position
+import nl.codestar.model.Position
 
 /**
   * Distance calculator in the WGS-84 system (World Geodetic System, version 1984)
@@ -16,13 +16,13 @@ trait DistanceCalculator {
 object Distance extends DistanceCalculator {
 
   private val AVERAGE_RADIUS_OF_EARTH_KM = 6371
-  private val AVERAGE_RADIUS_OF_EARTH_M = AVERAGE_RADIUS_OF_EARTH_KM * 1000
+  private val AVERAGE_RADIUS_OF_EARTH_M  = AVERAGE_RADIUS_OF_EARTH_KM * 1000
 
   private def haversineDistance(l1: Position, l2: Position): Double = {
     val latDistance = Math.toRadians(l1.latitude - l2.latitude)
     val lngDistance = Math.toRadians(l1.longitude - l2.longitude)
-    val sinLat = Math.sin(latDistance / 2)
-    val sinLng = Math.sin(lngDistance / 2)
+    val sinLat      = Math.sin(latDistance / 2)
+    val sinLng      = Math.sin(lngDistance / 2)
     val a = sinLat * sinLat +
       (Math.cos(Math.toRadians(l1.latitude)) *
         Math.cos(Math.toRadians(l2.latitude)) *
@@ -32,11 +32,11 @@ object Distance extends DistanceCalculator {
   }
 
   override def calculateInKilometers(l1: Position, l2: Position): Int = {
-    (AVERAGE_RADIUS_OF_EARTH_KM * haversineDistance(l1,l2)).toInt
+    (AVERAGE_RADIUS_OF_EARTH_KM * haversineDistance(l1, l2)).toInt
   }
 
   override def calculateInMeters(l1: Position, l2: Position): Int = {
-    (AVERAGE_RADIUS_OF_EARTH_M * haversineDistance(l1,l2)).toInt
+    (AVERAGE_RADIUS_OF_EARTH_M * haversineDistance(l1, l2)).toInt
   }
 
 }
